Mass separation of function into modules
