+++ express_server.js
@@ -1,163 +1,33 @@
-"use strict";
-
-const express = require("express");
+'use strict';
+require('dotenv').config();
+//Required npm packages --------------------------------------------------
+const express = require('express');
 const app = express();
-const PORT = process.env.PORT || 8080; // default port 8080
-const bodyParser = require("body-parser");
 const methodOverride = require('method-override');
-const MongoClient = require("mongodb").MongoClient;
+const bodyParser = require("body-parser");
+const MongoClient = require('mongodb').MongoClient;
 const MONGODB_URI = "mongodb://127.0.0.1:27017/url_shortener";
+const tAppFunc = require('./function_index');
+//Connection PORT to express server ---------------------------------------
+const PORT = process.env.PORT || 8080;
 
+app.use(bodyParser.urlencoded({ extended: true }));
 app.use(methodOverride('_method'));
-app.set("view engine", "ejs");
-app.use(bodyParser.urlencoded({ extended: false }));
-app.listen(PORT, () => {
- console.log(`Example app listening on port ${PORT}!`);
-});
+app.use(express.static('public'));
+app.set('view engine', 'ejs');
 
-
-
-//Generates a random string of 6 characters - when using insure it checks if string exists
-function generateRandomString(){
-  var text = "";
-  var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  for( var i=0; i < 6; i++ )
-    text += possible.charAt(Math.floor(Math.random() * possible.length));
-  return text;
-};
-
-var urlDatabase = {
-  "b2xVn2": "http://www.lighthouselabs.ca",
-  "9sm5xK": "http://www.google.com"
-};
-
-function insertURL(db, longURL, cb) {
-  let newURL = {
-    shortURL: generateRandomString(),
-    longURL: longURL
-  }
-  db.collection('urls').insertOne(newURL, (err, result) => {
-    if (err) {
-      return cb(err);
-    }
-    return cb(null, result);
-  });
-}
-
-function updateURL(db, shortURL, longURL, cb) {
-  db.collection('urls').updateOne(
-    { 'shortURL': shortURL },
-    {
-      $set: { 'longURL': longURL }
-    }, (err, result) => {
-    if (err) {
-      return cb(err);
-    }
-    return cb(null, result);
-  });
-}
-
-function deleteURL(db, shortURL, cb) {
-  db.collection('urls').remove({ 'shortURL': shortURL }, (err, result) => {
-    if (err) {
-      return cb(err);
-    }
-    return cb(null, result);
-  });
-}
-
-function getLongURL(db, shortURL, cb) {
-  let query = { 'shortURL': shortURL };
-  db.collection('urls').findOne(query, (err, result) => {
-    if (err) {
-      return cb(err);
-    }
-    return cb(null, result.longURL);
-  });
-}
-
-function getURLs(db, cb) {
-  db.collection('urls').find().toArray((err, result) => {
-    if (err) {
-      return cb(err);
-    }
-    return cb(null, result);
-  });
-}
-
-let dbInstance;
-
-MongoClient.connect(MONGODB_URI, (err, db) => {
+let db;
+//Initializing MongoDb Connection ------------------------------------------
+MongoClient.connect(MONGODB_URI, (err, dbInstance) => {
   if (err) {
-    console.log('Could not connect! Unexpected error. Details below.');
-    throw err;
+    console.log('Could not connect to the database! Details below.', err, 'The application will now exit.');
+    process.exit();
   }
-
-  dbInstance = db;
-
-  app.get("/urls/new", (req, res) => {
-   res.render("urls_new");
-  });
-
-  app.post("/urls", (req, res) => {
-    let longURL = req.body.longURL;
-    insertURL(db, longURL, (err, result) => {
-      res.redirect('/urls');
-    });
-  });
-
-  app.get('/urls', (req, res) => {
-    getURLs(db, (err, URLs) => {
-      let templateVars = { urls: URLs };
-      res.render('urls_index', templateVars);
-    });
-  });
-
-  app.get('/urls/:id', (req, res) => {
-    let shortURL = req.params.id;
-    getLongURL(db, shortURL, (err, longURL) => {
-      let templateVars = {
-        shortURL: shortURL,
-        longURL: longURL
-      };
-      res.render('urls_show', templateVars);
-    });
-  });
-
-  app.put('/urls/:id', (req, res) => {
-    let shortURL = req.params.id;
-    let longURL = req.body.url;
-    updateURL(db, shortURL, longURL, (err, result) => {
-      res.redirect('/urls');
-    });
-  });
-
-  app.delete('/urls/:id', (req, res) => {
-    let shortURL = req.params.id;
-    deleteURL(db, shortURL, (err, result) => {
-      res.redirect('/urls');
-    });
-  });
-
-  app.get("/u/:shortURL", (req, res) => {
-    let shortURL = req.params.shortURL;
-    getLongURL(db, shortURL, (err, longURL) => {
-      if (err === '404_Page_not_found') {
-        let templateVars = {
-          title: 'Not Found!',
-          shortURL: shortURL
-        };
-        res.status(404).render('404_Page_not_found', templateVars);
-      } else {
-        if(longURL.indexOf('http://') === -1) {
-          longURL = 'http://' + longURL;
-        }
-        res.status(301).redirect(longURL);
-      }
-    });
-  });
-
-  app.get('/urls.json', (req, res) => {
-   res.json(urlDatabase);
+  db = dbInstance;
+  require('./express_functions')(app, db);
+  require('./short')(app, db);
+  // Initiate server
+  app.listen(PORT, () => {
+    console.log(`Server listening on port ${PORT}!`);
   });
 });
