+++ views/404.ejs
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<header>
+  <% include ./partials/_header %>
+</header>
+
+<body>
+  <div>
+    Could not find a shortURL with the id: <%= shortURL %>
+  </div>
+</body>
+</html>
